gpwrap: wrap gphoto object creation for unique_ptr
wrap gphoto's construction functions to get both syntactic sugar and
unique pointer safety.

gphoto creates stuff with

  Foo* obj;
  if (gp_foo_new(&foo) < GP_OK)
    handle_error();

wow, so many lines of code. wrap it all and return a unique pointer so
that we're safe to forget deletion - the objects have a consistent
gp_foo_new and gp_foo_unref functions.

The GP_NEW_UNIQUE macro provides function names for exceptions in case
of allocation errors (as if it'd ever happen) and both creation and
deletion.